backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0
scikit-image==0.21.0
joblib==1.3.2
Pillow==10.0.0